Add CloseAndOpenWhileReceiving test
This test tries to trigger a race in socket handling.

Specifically it tries to make it so that one thread is about to block
on a recv(), while another one closes a socket and opens a new one
(getting the same file descriptor).

This results in the first thread blocking until the second socket is
closed, even though it was in theory waiting on the first one.