[NOT FOR MERGING] Introduce delay to trigger race
The socket handling code (and possibly also some other of the code
which uses wapi handles) is racy.

It does not always fail, but using appropriate delays it is possible
to cause threads to be scheduled in ways which show the misbehaviour.

On my system (MacOS X 10.9) I get the following behaviour:

0x112e84000 socket 6
0x112e84000 closed 6
0x112e84000 socket 6
0x113ef9000 sleeping on 6...
0x112e84000 closed 6
0x112e84000 socket 6
0x113ef9000 receiving on 6...
0x113ef9000 received on 6 (-1)
0x112e84000 closed 6

i.e. one thread gets ready to receive on a socket (sleeping), but in
the meantime another thread is able to close it and open a new one
with the same file descriptor.