Hack in K1OM support
The K1OM architecture shares most of the opcodes with AMD64, with the
notable exception of the SIMD instructions. In particular on K1OM the
XMM registers do not exist at all, which means that all of the
instructions that use them are invalid. The ABI expects that floating
point arguments are passed/returned in the ZMM registers, which are
currently not accessible in Mono.

Nontheless, simply disabling the saving of the XMM registers is
sufficient to get a partially working Mono, as long as the code that
is executed does not try to perform floating-point operations.